WalletScanLogic
